requirements.txt
blinker==1.6.2
cachelib==0.10.2
click==8.1.7
flake8==6.1.0
Flask==2.3.3
Flask-Session==0.5.0
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
mysql-connector==2.2.9
packaging==23.1
pycodestyle==2.11.0
pyflakes==3.1.0
style==1.1.0
update==0.0.1
Werkzeug==2.3.7
